requirements.txt
python-jobspy
flask
requests
gspread
google-auth
anthropic
openai
PyPDF2
python-docx
